Remove evil static state
Getting rid of virtually all static state in the service-a project.
The main intention of this is to be able to easily run multiple
pollers. Each poller is independent of the others, allowing
for simpler setup when multiple pollers are required.
Previously, there was a fair amount of ad-hoc static state
lying around the place.