Rearrange project structure and namespaces to get ready for nuget package creation
Time to bite-the-bullet and start putting a bit more order to this.
Intention is to make nuget packages available for others, for eductional purposes.
But, also want to keep code base in a single project for simplicity of
working through with others (and might make refactoring easier, lets see!)
So:
- assembly renaming.
- namespace fixing.
- work towards getting rid of carrier-pidgeon everywhere.